refactor: move console files out of wargames feat
The previous directory tree refactor caused two pages to be misplaced
into the wargames feature directory. Move them to the console feature
directory, which is how they are themed and linked from in the UI.
