+++ src/App.jsx
@@ -4,12 +4,12 @@ import RootLayout from "./layouts/RootLayout";
 import ConsoleLayout from "./features/console/layouts/ConsoleLayout";
 import WargamesLayout from "./features/wargames/layouts/WargamesLayout";
 import Home from "./pages/Home";
-import WargamesDashboard from "./features/wargames/pages/Wargames";
+import WargamesDashboard from "./features/console/pages/WargamesDashboard";
 import Models from "./features/console/pages/Models";
 import RedTeaming from "./features/console/pages/RedTeaming";
 import Dashboard from "./features/console/pages/Dashboard";
 import Tournaments from "./features/console/pages/Tournaments";
-import ChallengeDebug from "./features/wargames/pages/ChallengeDebug";
+import ChallengeDebug from "./features/console/pages/ChallengeDebug";
 import WargamesChallenge from "./features/wargames/pages/WargamesChallenge";
 import Callback from "./shared/components/auth/Callback";
 import { setupApiClient } from "./shared/lib/api-client";
