+++ src/features/console/pages/ChallengeDebug.jsx
